+++ jason/src/api/dart/utils/err.rs
@@ -5,12 +5,18 @@ use std::{borrow::Cow, ptr};
 use dart_sys::Dart_Handle;
 use derive_more::Into;
 use libc::c_char;
-use tracerr::{Trace, Traced};
 
 use crate::{
-    api::dart::{utils::string_into_c_str, DartValue},
+    api::{
+        dart::{utils::string_into_c_str, DartValue},
+        errors::{
+            EnumerateDevicesException, FormatException, InternalException,
+            LocalMediaInitException, LocalMediaInitExceptionKind,
+            MediaSettingsUpdateException, MediaStateTransitionException,
+            RpcClientException, RpcClientExceptionKind, StateError,
+        },
+    },
     platform,
-    room::ChangeMediaStateError,
 };
 
 /// Pointer to an extern function that returns a new Dart [`ArgumentError`] with
@@ -289,8 +295,6 @@ impl From<platform::Error> for DartError {
 
 /// Error returning by Rust when an unacceptable argument is passed to a
 /// function through FFI.
-///
-/// It can be converted into a [`DartError`] and passed to Dart.
 pub struct ArgumentError<T> {
     /// Invalid value of the argument.
     val: T,
@@ -332,86 +336,13 @@ impl<T: Into<DartValue>> From<ArgumentError<T>> for DartError {
     }
 }
 
-/// Error thrown when the operation wasn't allowed by the current state of the
-/// object.
-///
-/// It can be converted into a [`DartError`] and passed to Dart.
-pub struct StateError(Cow<'static, str>);
-
-impl StateError {
-    /// Creates a new [`StateError`] with the provided `message` describing the
-    /// problem.
-    #[inline]
-    #[must_use]
-    pub fn new<T: Into<Cow<'static, str>>>(message: T) -> Self {
-        Self(message.into())
-    }
-}
-
 impl From<StateError> for DartError {
     #[inline]
     fn from(err: StateError) -> Self {
         unsafe {
             Self::new(NEW_STATE_ERROR_CALLER.unwrap()(string_into_c_str(
-                err.0.into_owned(),
+                err.message(),
             )))
-        }
-    }
-}
-
-/// Possible error kinds of a [`LocalMediaInitException`].
-#[repr(u8)]
-pub enum LocalMediaInitExceptionKind {
-    /// Occurs if the [getUserMedia()][1] request failed.
-    ///
-    /// [1]: https://tinyurl.com/w3-streams#dom-mediadevices-getusermedia
-    GetUserMediaFailed,
-
-    /// Occurs if the [getDisplayMedia()][1] request failed.
-    ///
-    /// [1]: https://w3.org/TR/screen-capture/#dom-mediadevices-getdisplaymedia
-    GetDisplayMediaFailed,
-
-    /// Occurs when local track is [`ended`][1] right after [getUserMedia()][2]
-    /// or [getDisplayMedia()][3] request.
-    ///
-    /// [1]: https://tinyurl.com/w3-streams#idl-def-MediaStreamTrackState.ended
-    /// [2]: https://tinyurl.com/rnxcavf
-    /// [3]: https://w3.org/TR/screen-capture#dom-mediadevices-getdisplaymedia
-    LocalTrackIsEnded,
-}
-
-/// Exception thrown when accessing media devices.
-pub struct LocalMediaInitException {
-    /// Concrete error kind of this [`LocalMediaInitException`].
-    kind: LocalMediaInitExceptionKind,
-
-    /// Error message describing the problem.
-    message: Cow<'static, str>,
-
-    /// [`platform::Error`] that caused this [`LocalMediaInitException`].
-    cause: Option<platform::Error>,
-
-    /// Stacktrace of this [`LocalMediaInitException`].
-    trace: Trace,
-}
-
-impl LocalMediaInitException {
-    /// Creates a new [`LocalMediaInitException`] from the provided error
-    /// `kind`, `message`, optional `cause` and `trace`.
-    #[inline]
-    #[must_use]
-    pub fn new<M: Into<Cow<'static, str>>>(
-        kind: LocalMediaInitExceptionKind,
-        message: M,
-        cause: Option<platform::Error>,
-        trace: Trace,
-    ) -> Self {
-        Self {
-            kind,
-            message: message.into(),
-            cause,
-            trace,
         }
     }
 }
@@ -421,31 +352,12 @@ impl From<LocalMediaInitException> for DartError {
     fn from(err: LocalMediaInitException) -> Self {
         unsafe {
             Self::new(NEW_LOCAL_MEDIA_INIT_EXCEPTION_CALLER.unwrap()(
-                err.kind,
-                string_into_c_str(err.message.into_owned()),
-                err.cause.map(DartError::from).into(),
-                string_into_c_str(err.trace.to_string()),
+                err.kind(),
+                string_into_c_str(err.message()),
+                err.cause().map(DartError::from).into(),
+                string_into_c_str(err.trace()),
             ))
         }
-    }
-}
-
-/// Exception thrown when cannot get info of available media devices.
-pub struct EnumerateDevicesException {
-    /// [`platform::Error`] that caused this [`EnumerateDevicesException`].
-    cause: platform::Error,
-
-    /// Stacktrace of this [`EnumerateDevicesException`].
-    trace: Trace,
-}
-
-impl EnumerateDevicesException {
-    /// Creates a new [`EnumerateDevicesException`] from the provided error
-    /// `cause` and `trace`.
-    #[inline]
-    #[must_use]
-    pub fn new(cause: platform::Error, trace: Trace) -> Self {
-        Self { cause, trace }
     }
 }
 
@@ -454,26 +366,10 @@ impl From<EnumerateDevicesException> for DartError {
     fn from(err: EnumerateDevicesException) -> Self {
         unsafe {
             Self::new(NEW_ENUMERATE_DEVICES_EXCEPTION_CALLER.unwrap()(
-                err.cause.into(),
-                string_into_c_str(err.trace.to_string()),
+                err.cause().into(),
+                string_into_c_str(err.trace()),
             ))
         }
-    }
-}
-
-/// Exception thrown when a string or some other data doesn't have an expected
-/// format and cannot be parsed or processed.
-///
-/// It can be converted into a [`DartError`] and passed to Dart.
-pub struct FormatException(Cow<'static, str>);
-
-impl FormatException {
-    /// Creates a new [`FormatException`] with the provided `message` describing
-    /// the problem.
-    #[inline]
-    #[must_use]
-    pub fn new<T: Into<Cow<'static, str>>>(message: T) -> Self {
-        Self(message.into())
     }
 }
 
@@ -482,63 +378,8 @@ impl From<FormatException> for DartError {
     fn from(err: FormatException) -> Self {
         unsafe {
             Self::new(NEW_FORMAT_EXCEPTION_CALLER.unwrap()(string_into_c_str(
-                err.0.into_owned(),
+                err.message(),
             )))
-        }
-    }
-}
-
-/// Possible error kinds of a [`RpcClientException`].
-#[repr(u8)]
-pub enum RpcClientExceptionKind {
-    /// Connection with a server was lost.
-    ///
-    /// This usually means that some transport error occurred, so a client can
-    /// continue performing reconnecting attempts.
-    ConnectionLost,
-
-    /// Could not authorize an RPC session.
-    ///
-    /// This usually means that authentication data a client provides is
-    /// obsolete.
-    AuthorizationFailed,
-
-    /// RPC session has been finished. This is a terminal state.
-    SessionFinished,
-}
-
-/// Exceptions thrown from an RPC client that implements messaging with media
-/// server.
-pub struct RpcClientException {
-    /// Concrete error kind of this [`RpcClientException`].
-    kind: RpcClientExceptionKind,
-
-    /// Error message describing the problem.
-    message: Cow<'static, str>,
-
-    /// [`platform::Error`] that caused this [`RpcClientException`].
-    cause: Option<platform::Error>,
-
-    /// Stacktrace of this [`RpcClientException`].
-    trace: Trace,
-}
-
-impl RpcClientException {
-    /// Creates a new [`RpcClientException`] from the provided error `kind`,
-    /// `message`, optional `cause` and `trace`.
-    #[inline]
-    #[must_use]
-    pub fn new<M: Into<Cow<'static, str>>>(
-        kind: RpcClientExceptionKind,
-        message: M,
-        cause: Option<platform::Error>,
-        trace: Trace,
-    ) -> Self {
-        Self {
-            kind,
-            message: message.into(),
-            cause,
-            trace,
         }
     }
 }
@@ -548,34 +389,11 @@ impl From<RpcClientException> for DartError {
     fn from(err: RpcClientException) -> Self {
         unsafe {
             Self::new(NEW_RPC_CLIENT_EXCEPTION_CALLER.unwrap()(
-                err.kind,
-                string_into_c_str(err.message.into_owned()),
-                err.cause.map(DartError::from).into(),
-                string_into_c_str(err.trace.to_string()),
+                err.kind(),
+                string_into_c_str(err.message()),
+                err.cause().map(DartError::from).into(),
+                string_into_c_str(err.trace()),
             ))
-        }
-    }
-}
-
-/// Exception thrown when the requested media state transition could not be
-/// performed.
-pub struct MediaStateTransitionException {
-    /// Error message describing the problem.
-    message: Cow<'static, str>,
-
-    /// Stacktrace of this [`MediaStateTransitionException`].
-    trace: Trace,
-}
-
-impl MediaStateTransitionException {
-    /// Creates a new [`MediaStateTransitionException`] from the provided error
-    /// `message` and `trace`.
-    #[inline]
-    #[must_use]
-    pub fn new<T: Into<Cow<'static, str>>>(message: T, trace: Trace) -> Self {
-        Self {
-            message: message.into(),
-            trace,
         }
     }
 }
@@ -585,44 +403,9 @@ impl From<MediaStateTransitionException> for DartError {
     fn from(err: MediaStateTransitionException) -> Self {
         unsafe {
             Self::new(NEW_MEDIA_STATE_TRANSITION_EXCEPTION_CALLER.unwrap()(
-                string_into_c_str(err.message.into_owned()),
-                string_into_c_str(err.trace.to_string()),
+                string_into_c_str(err.message()),
+                string_into_c_str(err.trace()),
             ))
-        }
-    }
-}
-
-/// Jason's internal exception.
-///
-/// This is either a programmatic error or some unexpected platform component
-/// failure that cannot be handled in any way.
-///
-/// It can be converted into a [`DartError`] and passed to Dart.
-pub struct InternalException {
-    /// Error message describing the problem.
-    message: Cow<'static, str>,
-
-    /// [`platform::Error`] that caused this [`RpcClientException`].
-    cause: Option<platform::Error>,
-
-    /// Stacktrace of this [`InternalException`].
-    trace: Trace,
-}
-
-impl InternalException {
-    /// Creates a new [`InternalException`] from the provided error `message`,
-    /// `trace` and an optional `cause`.
-    #[inline]
-    #[must_use]
-    pub fn new<T: Into<Cow<'static, str>>>(
-        message: T,
-        cause: Option<platform::Error>,
-        trace: Trace,
-    ) -> Self {
-        Self {
-            message: message.into(),
-            trace,
-            cause,
         }
     }
 }
@@ -632,46 +415,10 @@ impl From<InternalException> for DartError {
     fn from(err: InternalException) -> Self {
         unsafe {
             Self::new(NEW_INTERNAL_EXCEPTION_CALLER.unwrap()(
-                string_into_c_str(err.message.into_owned()),
-                err.cause.map(DartError::from).into(),
-                string_into_c_str(err.trace.to_string()),
+                string_into_c_str(err.message()),
+                err.cause().map(DartError::from).into(),
+                string_into_c_str(err.trace()),
             ))
-        }
-    }
-}
-
-/// Errors occurring in [`RoomHandle::set_local_media_settings()`][1] method.
-///
-/// It can be converted into a [`DartError`] and passed to Dart.
-///
-/// [1]: crate::api::RoomHandle::set_local_media_settings
-pub struct MediaSettingsUpdateException {
-    /// Error message describing the problem.
-    message: Cow<'static, str>,
-
-    /// Original [`ChangeMediaStateError`] that was encountered while updating
-    /// local media settings.
-    cause: Traced<ChangeMediaStateError>,
-
-    /// Whether media settings were successfully rolled back after new settings
-    /// application failed.
-    rolled_back: bool,
-}
-
-impl MediaSettingsUpdateException {
-    /// Creates a new [`MediaSettingsUpdateException`] from the provided error
-    /// `message`, `cause` and `rolled_back` property.
-    #[inline]
-    #[must_use]
-    pub fn new<T: Into<Cow<'static, str>>>(
-        message: T,
-        cause: Traced<ChangeMediaStateError>,
-        rolled_back: bool,
-    ) -> Self {
-        Self {
-            message: message.into(),
-            rolled_back,
-            cause,
         }
     }
 }
@@ -681,9 +428,9 @@ impl From<MediaSettingsUpdateException> for DartError {
     fn from(err: MediaSettingsUpdateException) -> Self {
         unsafe {
             Self::new(NEW_MEDIA_SETTINGS_UPDATE_EXCEPTION_CALLER.unwrap()(
-                string_into_c_str(err.message.into_owned()),
-                err.cause.into(),
-                err.rolled_back as u8,
+                string_into_c_str(err.message()),
+                err.cause(),
+                err.rolled_back() as u8,
             ))
         }
     }
