+++ jason/src/api/mod.rs
@@ -1,5 +1,7 @@
 //! External [`Jason`] API.
 
+mod errors;
+
 cfg_if::cfg_if! {
     if #[cfg(target_os = "android")] {
         mod dart;
