+++ jason/src/platform/transport.rs
@@ -65,7 +65,7 @@ pub trait RpcTransport {
 #[js(error = "platform::Error")]
 pub enum TransportError {
     /// Error encountered when trying to establish connection.
-    #[display(fmt = "Failed to create WebSocket: {}", _0)]
+    #[display(fmt = "Failed to create WebSocket: {:?}", _0)]
     CreateSocket(platform::Error),
 
     /// Connection was closed before becoming active.
@@ -85,7 +85,7 @@ pub enum TransportError {
     MessageNotString,
 
     /// Occurs when a message cannot be send to server.
-    #[display(fmt = "Failed to send message: {}", _0)]
+    #[display(fmt = "Failed to send message: {:?}", _0)]
     SendMessage(platform::Error),
 
     /// Occurs when message is sent to a closed socket.
